requirements.txt
torch
torchvision
torchaudio
torchmetrics
pandas
numpy
scikit-learn
PyYAML
tqdm
argparse
-e .